40-17G #CONCEALED BY CIK FILER 93715 AND CIK FILER 1515623
(b) Amount of penalty -

(1) In general. The amount of the accuracy-related penalty is 20 percent of the portion of an underpayment
of tax required to be shown on a return that is attributable to any of the types of misconduct listed in paragraphs
(a)(1) through (a)(3) of this section, except as provided in paragraph (b)(2) of this section.

(2) Increase in penalty for gross valuation misstatement.
In the case of a gross valuation misstatement, as defined in section 6662(h)(2) and § 1.6662-5(e)(2),
the amount of the accuracy-related penalty is 40 percent of the portion of an underpayment of tax required
to be shown on a return that is attributable to the gross valuation misstatement, provided the applicable
dollar limitation set forth in section 6662(e)(2) is satisfied.



§ 1.6662-5 Substantial and gross valuation misstatements under chapter 1.
(a) In general. If any portion of an underpayment, as defined in section 6664(a) and
§ 1.6664-2, of any income tax imposed under chapter 1 of subtitle A of the Code that is
required to be shown on a return is attributable to a substantial valuation misstatement
under chapter 1 (“substantial valuation misstatement”), there is added to the tax an amount equal to 20 percent of such portion.

Section 6662(h) increases the penalty to 40 percent in the case of a gross valuation misstatement under chapter 1
(“gross valuation misstatement”).
No penalty under section 6662(b)(3) is imposed, however, on a portion of an underpayment
that is attributable to a substantial or gross valuation misstatement unless the aggregate
of all portions of the underpayment attributable to substantial or gross valuation misstatements
exceeds the applicable dollar limitation ($5,000 or $10,000), as provided in section 6662(e)(2) and paragraphs
(b) and
(f)(2) of this section.
This penalty also does not apply to the extent that the reasonable cause and good faith exception to this penalty
set forth in § 1.6664-4 applies. There is no disclosure exception to this penalty.

(b) Dollar limitation.
No penalty may be imposed under section 6662(b)(3) for a taxable year unless

the portion of the underpayment for that year that is attributable to substantial or gross valuation misstatements exceeds $5,000
($10,000 in the case of a corporation other than an S corporation (as defined in section 1361(a)(1));

or a personal holding company (as defined in section 542)).

This limitation is applied separately to each taxable year for which there is a substantial or gross valuation misstatement.



(e) Definitions -

(1) Substantial valuation misstatement.
There is a substantial valuation misstatement if the value or adjusted basis of any property claimed on a
return of tax imposed under chapter 1 is 200 percent or more of the correct amount.

(2) Gross valuation misstatement.
There is a gross valuation misstatement if the value or adjusted basis of any property claimed on a return of tax
imposed under chapter 1 is 400 percent or more of the correct amount.

(3) Property.
For purposes of this section, the term “property” refers to both tangible and intangible property.
Tangible property includes property such as land, buildings, fixtures and inventory.
Intangible property includes property such as goodwill, covenants not to compete, leaseholds, patents,
contract rights, debts and choses in action.


(f) Multiple valuation misstatements on a return -

(1) Determination of whether valuation misstatements are substantial or gross.
The determination of whether there is a substantial or gross valuation misstatement on a return is
made on a property-by-property basis. Assume, for example, that property A has a value of 60 but a taxpayer claims a value of 110,
and that property B has a value of 40 but the taxpayer claims a value of 100.

Because the claimed and correct values are compared on a property-by-property basis,
there is a substantial valuation misstatement with respect to property B, but not with respect to property A,
even though the claimed values (210) are 200 percent or more of the correct values (100) when compared on an aggregate basis.

(2) Application of dollar limitation. For purposes of applying the dollar limitation set forth in section 6662(e)(2),
the determination of the portion of an underpayment that is attributable to a substantial or gross valuation misstatement
is made by aggregating all portions of the underpayment attributable to substantial or gross valuation misstatements.

Assume, for example, that the value claimed for property C on a return is 250 percent of the correct value, and that the value claimed for property D on the return is 400 percent of the correct value.
Because the portions of an underpayment that are attributable to a substantial or gross valuation misstatement on a return are aggregated in applying the dollar limitation, the dollar limitation is satisfied if the portion of the underpayment that is attributable to the misstatement of the value of property C, when aggregated with the portion of the underpayment that is attributable to the misstatement of the value of property D, exceeds $5,000 ($10,000 in the case of most corporations).

(g) Property with a value or adjusted basis of zero.

The value or adjusted basis claimed on a return of any property with a correct value or adjusted basis of zero is considered to be 400 percent or more of the correct amount. There is a gross valuation misstatement with respect to such property, therefore, and the applicable penalty rate is 40 percent.

I ANNEXED THESE - JUST IN CASE.
RECEIVED NYSCEF: 08/01/2020

https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=2xkNo2v2/wHD14V7b_PLUS_2njQ==

HERE'S THE INVOICE WITH JUNE PAID. INCLUDING LEGAL FEES AND ANY DAMAGES.

https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=vd54Sn1RLwYlVxoKhGh0xg==